feature: initial breakdown of prototype
