+++ server/router.js
@@ -2,5 +2,5 @@
 // Do route dispatch here
 
 module.exports = app => {
-  app.use('/', require('./routes/index'));
+  app.use('/', require('./routes/applications'));
 };
