+++ server/views/_layouts/default.njk
@@ -17,13 +17,27 @@
     {# <script>if (typeof window.GOVUK === 'undefined') document.body.className = document.body.className.replace('js-enabled', '');</script> #}
   </head>
 
-  <body class="govuk-template__body">
+  <body class="govuk-template__body js-enabled">
+
+    <script>document.body.className = ((document.body.className) ? document.body.className + ' js-enabled' : 'js-enabled');</script>
 
     {% include "_partials/__header.njk" %}
 
-    <div class="govuk-width-container">
-      {% block main_content %}
-      {% endblock %}
+    <div class="govuk-width-container ">
+
+      <main class="govuk-main-wrapper " id="main-content" role="main">
+
+        <div class="govuk-grid-row">
+          <div class="govuk-grid-column-one-third">
+            {% include "_partials/nav_bar.njk" %}
+          </div>
+          <div class="govuk-grid-column-two-thirds">
+            {% block main_content %}
+            {% endblock %}
+          </div>
+        </div>
+
+      </main>
     </div>
 
     {% include "_partials/__footer.njk" %}
