+++ server/views/_partials/__header.njk
@@ -4,4 +4,38 @@
     {% include "_partials/banner.njk" %}
   </div>
 </header>
-{% include "_partials/phase_banner.njk" %}
+<div class="govuk-width-container">
+    <div class="govuk-phase-banner">
+      <p class="govuk-phase-banner__content">
+        <strong class="govuk-tag govuk-phase-banner__content__tag">
+          beta
+        </strong>
+        <span class="govuk-phase-banner__text">
+          This is a new service – your <a class="govuk-link" href="#">feedback</a> will help us to improve it.
+        </span>
+      </p>
+    </div>
+
+    <div class="js-toggle-nav" id="global-nav">
+      <div id="user_bar">
+        <nav role="navigation" class="content">
+          <ul id="navigation">
+
+
+              <li id="signed-in-user" class="user">
+                demo@ch.gov.uk
+              </li>
+              <li>
+                <a id="user-signout" href="/">Sign out</a>
+              </li>
+
+            <li id="signed-in-user" class="user">
+              <a href="https://forum.aws.chdev.org">Developer hub forum</a>
+            </li>
+          </ul>
+
+        </nav>
+      </div>
+    </div>
+</div>
+
