+++ server/views/_partials/nav_bar.njk
@@ -1,0 +1,33 @@
+      <nav class="app-subnav app-hide-mobile">
+
+          <h4 class="app-subnav__theme">Manage applications</h4>
+          <ul class="app-subnav__section">
+            <li class="app-subnav__section-item app-subnav__section-item--current">
+              <a class="app-subnav__link govuk-link " href="manage-applications">
+                View all applications
+              </a>
+            </li>
+            <li class="app-subnav__section-item">
+              <a class="app-subnav__link govuk-link " href="../register">
+                Add an application
+              </a>
+            </li>
+          </ul>
+
+        <h4 class="app-subnav__theme">General documentation</h4>
+        <ul class="app-subnav__section">
+          <li class="app-subnav__section-item ">
+            <a class="app-subnav__link govuk-link" href="/general-documentation/get-started.html">Get started</a>
+          </li>
+          <li class="app-subnav__section-item ">
+            <a class="app-subnav__link govuk-link" href="/general-documentation/rest-api-overview.html">Companies House REST API overview</a>
+          </li>
+          <li class="app-subnav__section-item ">
+            <a class="app-subnav__link govuk-link" href="/general-documentation/add-an-api-key.html">How to add an API key</a>
+          </li>
+          <li class="app-subnav__section-item ">
+            <a class="app-subnav__link govuk-link" href="/general-documentation/developer-guidelines.html">Developer guidelines</a>
+          </li>
+        </ul>
+
+      </nav>
