+++ server/views/applications/index.njk
